some things

+++ cap_lib/include/cap/string.h
@@ -11,7 +11,7 @@ struct String {
 
 #define str(s) ((String){.data = ((char*)(u64)(s)), .length = sizeof((s)) - 1})
 
-#define str_info(s) s.length, s.data
+#define str_info(s) (int)(s).length, (s).data
 
 bool string_equal(String a, String b);
